:bug: Fix the progress ending at 59:59
The duration is at 00:00 at the beginning and the end. Maybe make it to display the duration of the audio file.

+++ soundwave/src/main/java/fr/haran/soundwave/controller/PlayerListener.kt
@@ -5,5 +5,5 @@ interface PlayerListener {
     fun onPause(playerController: PlayerController)
     fun onPlay(playerController: PlayerController)
     fun onPrepared(playerController: PlayerController)
-    fun onDurationProgress(playerController: PlayerController, duration: Int, currentTimeStamp: Long)
+    fun onDurationProgress(playerController: PlayerController, duration: Int, currentTimeStamp: Int)
 }
